+++ Core/wbDefinitionsFNV.pas
@@ -9834,7 +9834,7 @@ begin
       ]),
       wbRArrayS('Linked Rooms',
         wbFormIDCk(XLRM, 'Linked Room', [REFR])
-      )
+      ).SetCountPath('XRMR\Linked Rooms Count')
     ], []),
 
     wbSizePosRot(XOCP, 'Occlusion Plane Data'),