+++ Core/wbDefinitionsFO3.pas
@@ -9290,7 +9290,7 @@ begin
       ]),
       wbRArrayS('Linked Rooms',
         wbFormIDCk(XLRM, 'Linked Room', [REFR])
-      )
+      ).SetCountPath('XRMR\Linked Rooms Count')
     ], []),
 
     wbSizePosRot(XOCP, 'Occlusion Plane Data'),