+++ Core/wbImplementation.pas
@@ -381,6 +381,7 @@ type
     procedure MergeStorageInternal(var aBasePtr: Pointer; aEndPtr: Pointer); virtual;
     procedure InformStorage(var aBasePtr: Pointer; aEndPtr: Pointer); virtual;
     procedure Remove; virtual;
+    procedure BeforeActualRemove; virtual;
     function CanContainFormIDs: Boolean; virtual;
     function AddIfMissing(const aElement: IwbElement; aAsNew, aDeepCopy : Boolean; const aPrefixRemove, aSuffixRemove, aPrefix, aSuffix: string; aAllowOverwrite: Boolean): IwbElement;
     function AddIfMissingInternal(const aElement: IwbElement; aAsNew, aDeepCopy : Boolean; const aPrefixRemove, aSuffixRemove, aPrefix, aSuffix: string; aAllowOverwrite: Boolean): IwbElement; virtual;
@@ -1821,6 +1822,9 @@ type
     procedure DoInit(aNeedSorted: Boolean); override;
 
     function Add(const aName: string; aSilent: Boolean): IwbElement; override;
+
+    procedure BeforeActualRemove; override;
+    procedure DoAfterSet(const aOldValue, aNewValue: Variant); override;
 
     function GetValue: string; override;
     function GetCheck: string; override;
@@ -16113,6 +16117,11 @@ begin
   Result := TargetValueDef.Assign(Self, aIndex, aElement, aOnlySK);;
 end;
 
+procedure TwbElement.BeforeActualRemove;
+begin
+  {can be overriden}
+end;
+
 procedure TwbElement.BeforeDestruction;
 begin
   if esConstructionComplete in eStates then
@@ -17124,6 +17133,7 @@ begin
     try
       SetModified(True);
       InvalidateParentStorage;
+      BeforeActualRemove;
       lContainer.RemoveElement(SelfRef);
     finally
       lContainer.EndUpdate;
@@ -17647,6 +17657,34 @@ begin
   end;
 end;
 
+procedure TwbSubRecordArray.BeforeActualRemove;
+var
+  CountPath      : string;
+  Container      : IwbContainerElementRef;
+  CounterElement : IwbElement;
+
+  SelfRef    : IwbContainerElementRef;
+begin
+  SelfRef := Self;
+
+  inherited;
+
+  CountPath := arcDef.CountPath;
+
+  if CountPath = '' then
+    Exit;
+
+  Container := GetContainer as IwbContainerElementRef;
+  if not Assigned(Container) then
+    Exit;
+
+  CounterElement := Container.ElementByPath[CountPath];
+  if not Assigned(CounterElement) then
+    Exit;
+
+  CounterElement.NativeValue := 0;
+end;
+
 function TwbSubRecordArray.CanAssignInternal(aIndex: Integer; const aElement: IwbElement; aCheckDontShow: Boolean): Boolean;
 begin
   Result := False;
@@ -17715,6 +17753,34 @@ begin
     SetModified(True);
     InvalidateStorage;
   end;
+end;
+
+procedure TwbSubRecordArray.DoAfterSet(const aOldValue, aNewValue: Variant);
+var
+  CountPath      : string;
+  Container      : IwbContainerElementRef;
+  CounterElement : IwbElement;
+
+  SelfRef    : IwbContainerElementRef;
+begin
+  SelfRef := Self;
+
+  inherited;
+
+  CountPath := arcDef.CountPath;
+
+  if CountPath = '' then
+    Exit;
+
+  Container := GetContainer as IwbContainerElementRef;
+  if not Assigned(Container) then
+    Exit;
+
+  CounterElement := Container.ElementByPath[CountPath];
+  if not Assigned(CounterElement) then
+    Exit;
+
+  CounterElement.NativeValue := GetElementCount;
 end;
 
 procedure TwbSubRecordArray.DoInit(aNeedSorted: Boolean);