+++ Core/wbInterface.pas
@@ -2200,9 +2200,13 @@ type
     ['{67943BAC-B558-4112-8DBC-C94A44E0B1D1}']
     function GetElement: IwbRecordMemberDef;
     function GetSorted(const aContainer: IwbContainer): Boolean;
+    function GetCountPath: string;
+
+    function SetCountPath(const aValue: string): IwbSubRecordArrayDef;
 
     property Element: IwbRecordMemberDef read GetElement;
     property Sorted[const aContainer: IwbContainer]: Boolean read GetSorted;
+    property CountPath: string read GetCountPath;
   end;
 
   IwbSubRecordStructDef = interface(IwbRecordMemberDef)
@@ -5317,6 +5321,7 @@ type
     sraElement  : IwbRecordMemberDef;
     sraSorted   : Boolean;
     sraIsSorted : TwbIsSortedCallback;
+    sraCountPath: string;
   public
     constructor Clone(const aSource: TwbDef); override;
     constructor Create(aPriority  : TwbConflictPriority; aRequired: Boolean;
@@ -5328,6 +5333,7 @@ type
                        aDontShow  : TwbDontShowCallback;
                        aIsSorted  : TwbIsSortedCallback;
                        aGetCP     : TwbGetConflictPriority); reintroduce;
+    procedure AfterClone(const aSource: TwbDef); override;
 
     {---IwbDef---}
     function GetDefType: TwbDefType; override;
@@ -5355,6 +5361,9 @@ type
     {---IwbSubRecordArrayDef---}
     function GetElement: IwbRecordMemberDef;
     function GetSorted(const aContainer: IwbContainer): Boolean;
+    function GetCountPath: string;
+
+    function SetCountPath(const aValue: string): IwbSubRecordArrayDef;
   end;
 
   TwbSubRecordStructDef = class(TwbRecordMemberDef, IwbSubRecordStructDef, IwbRecordDef)
@@ -9922,6 +9931,14 @@ end;
 
 { TwbSubRecordArrayDef }
 
+procedure TwbSubRecordArrayDef.AfterClone(const aSource: TwbDef);
+begin
+  inherited AfterClone(aSource);
+  with aSource as TwbSubRecordArrayDef do begin
+    Self.sraCountPath := sraCountPath;
+  end;
+end;
+
 procedure TwbSubRecordArrayDef.AfterLoad(const aElement: IwbElement);
 var
   Container: IwbContainerElementRef;
@@ -9988,6 +10005,11 @@ begin
   Result := sraElement;
 end;
 
+function TwbSubRecordArrayDef.GetCountPath: string;
+begin
+  Result := sraCountPath;
+end;
+
 function TwbSubRecordArrayDef.GetDefaultSignature: TwbSignature;
 begin
   Result := sraElement.GetDefaultSignature;
@@ -10028,6 +10050,12 @@ begin
   end;
 
   defReported := True;
+end;
+
+function TwbSubRecordArrayDef.SetCountPath(const aValue: string): IwbSubRecordArrayDef;
+begin
+  Result := Self;
+  sraCountPath := aValue;
 end;
 
 function TwbSubRecordArrayDef.ToSummaryInternal(aDepth: Integer; const aElement: IwbElement): string;