+++ whatsnew.md
@@ -12,6 +12,7 @@
 * #863 - FO4: DefaultObject forms (DFOB) should override via EditorID
 * #871 - Updates for Static Collections [FO4] 
 * #875 - TES4Edit 4.0.3 not able to copy Magic Effect Records
+* #919 - For REFR, Linked Rooms Count needs to be updated whenever the Linked Rooms are changed
 * (reported on Discord) - .esl extensions don't always force loading as ESL correctly when the module is not ESL flagged
 * (reported on Discord) - [TES4] EDID and ESCE incorrectly report "Expected 4 bytes but found 0" when running "Check for Errors" on MGEF
 * (found by developer) - Need to force InternalEdit for AfterSet handlers